[Re-]Format account address at DB layer
Flow Wallet API utilizes Account addresses as string across codebase.
When handling the addresses, the code prefers the format with `0x`
prefix while `flow.Address.Hex()` yields one without the prefix. Both
formats can be used theoretically and e.g. `flow.HexToAddress()` accepts
both formats the same.

Nevertheless, the address is stored as a string in the DB and this
introduces a problem if the addresses aren't formatted in a unified way.

If account has been persisted with `0x` prefix but later queried
without, there won't be a match.

This change adds formatting at DB layer to ensure unified format for
storing and querying.